+++ README.md
@@ -6,4 +6,9 @@ java坦克大战小游戏和中间的某些练习
 使用WSAD键进行上下左右移动坦克。
 按下J键发射子弹。
 
+## 目录结构
+
+* demo 坦克第一代原型
+* game 坦克第二代
+
 ![image-20230612150105023](https://wrxinyue.oss-cn-hongkong.aliyuncs.com/img/202306121501099.png)
