+++ doc/思路.md
@@ -1,0 +1,19 @@
+## 实现按下j键，我们的坦克发射子弹思路
+ > 1. 发射一颗子弹后，就相当于启动了一个线程
+ > 2. ho 类有子弹的对象，当按下j时，外面就启动一个发送行为（线程），让子弹不停的移动，形成一个射击效果
+ > 3. 让ht不停的重绘，才能出现子弹移动的效果
+ > 4. 当子弹移动到边界的时候，就销毁(退出线程)
+
+## 线程应用到坦克大战，让敌方发射子弹
+> 1. 让敌人的坦克也能发射子弹（可以有多颗子弹）
+> 2. 当我方坦克击中敌人坦克时，敌人的坦克就消失，可以做出爆炸的效果
+> 3. 让敌人的坦克可以自由随机移动
+> 4. 控制我方坦克和敌人坦克在规定的范围移动
+
+* 让敌人坦克也能发射子弹（多颗子弹）思路
+ >1. 在敌人坦克类，使用Vector 保存多个Shot，Vector是线程安全的所以使用在多线程 
+ >2. 当每创建创建敌人坦克时，就该给敌人坦克初始化一个shot对象，同时启动Shot
+ >3. 在绘制敌人坦克时，需要遍历敌人坦克对象的Vector，绘制所有的子弹，当子弹isLive == false时，就从Vector移出
+* 让敌人的坦克可以自由随机上下左右移动，因此需要将敌人坦克当作线程使用
+* 我们需要将EnemyTan 实现线程接口
+* 在run方法写上相应的业务代码
