--- src/cm/tk/ziji.java
@@ -1,7 +1,0 @@
-package cm.tk;
-//自己的坦克
-public class ziji extends tk{
-    public ziji(int x, int y) {
-        super(x, y);
-    }
-}
