+++ src/demo/ziji.java
@@ -1,0 +1,7 @@
+package demo;
+//自己的坦克
+public class ziji extends tk{
+    public ziji(int x, int y) {
+        super(x, y);
+    }
+}
